changed to float

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -9,7 +9,7 @@ public class MyMathCalc {
         return a - b;
     }
 
-    public int div(int a, int b){
+    public float div(int a, int b){
         return a / b;
 
     }
